\name{sumR}
\alias{sumR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Add the integer value in a string
}
\description{
sum all the interger values x
}
\usage{
sumR(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A numeric interger value x
}
}
\details{
Calculate the summation of an integer
}
\value{
Input integer vector 
}
\references{
Md. Joynal Abedin
}
\author{
Md. Shahjaman
}

